+++ src/pytest_randomly/__init__.py
@@ -51,6 +51,14 @@ try:
 except ImportError:  # pragma: no cover
     have_faker = False
 
+# model_bakery
+try:
+    from model_bakery.random_gen import baker_random
+
+    have_model_bakery = True
+except ImportError:  # pragma: no cover
+    have_model_bakery = False
+
 # numpy
 try:
     from numpy import random as np_random
@@ -160,6 +168,9 @@ def _reseed(config: Config, offset: int = 0) -> int:
 
     if have_faker:  # pragma: no branch
         faker_random.setstate(random_states[seed])
+
+    if have_model_bakery:  # pragma: no branch
+        baker_random.setstate(random_states[seed])
 
     if have_numpy:  # pragma: no branch
         numpy_seed = _truncate_seed_for_numpy(seed)