+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Reset the random state for `Model Bakery <https://model-bakery.readthedocs.io/en/latest/>`__.
+
 3.13.0 (2023-07-10)
 -------------------